+refactored structure of folders +started implementing Jetpack_Navigation
