+++ .idea/misc.xml
@@ -6,6 +6,7 @@
         <entry key="app/src/main/res/layout/activity_breeds.xml" value="1.0" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.26992753623188404" />
         <entry key="app/src/main/res/layout/fragment_breed_photos.xml" value="0.27309782608695654" />
+        <entry key="app/src/main/res/layout/fragment_breeds_list.xml" value="0.28125" />
         <entry key="app/src/main/res/layout/fragment_main.xml" value="0.3115942028985507" />
         <entry key="app/src/main/res/layout/fragment_start.xml" value="0.3244176013805004" />
       </map>
