+++ app/src/main/java/com/example/dogsapp/MainActivity.kt
@@ -9,10 +9,5 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        if (savedInstanceState == null) {
-            supportFragmentManager.beginTransaction()
-                .replace(R.id.container, StartFragment.newInstance())
-                .commitNow()
-        }
     }
 }
