--- app/src/main/java/com/example/dogsapp/breedPhotos.kt
@@ -1,17 +1,0 @@
-package com.example.dogsapp
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-
-/**
- * A simple [Fragment] subclass.
- * Use the [breedPhotos.newInstance] factory method to
- * create an instance of this fragment.
- */
-class breedPhotos : Fragment() {
-
-}
