+++ app/src/main/java/com/example/dogsapp/breeds/BreedsListFragment.kt
@@ -1,0 +1,8 @@
+package com.example.dogsapp.breeds
+
+
+import androidx.fragment.app.Fragment
+
+class BreedsListFragment : Fragment() {
+
+}
