+++ app/src/main/res/layout/fragment_breeds_list.xml
@@ -3,10 +3,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".breeds.BreedPhotos">
+    tools:context="com.example.dogsapp.breeds.BreedsListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/breedPhotosRv"
+        android:id="@+id/breedsListRv"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 </FrameLayout>
